--- src/app/[city]/loader.tsx
@@ -1,5 +1,0 @@
-import { Loader } from '@/components/Loader/Loader'
-
-export default function Loading() {
-  return <Loader />
-}
