+++ src/app/loading.tsx
@@ -1,0 +1,5 @@
+import { Loader } from '@/components/Loader/Loader'
+
+const Loading = () => <Loader />
+
+export default Loading
