+++ src/components/Header/Header.tsx
@@ -1,3 +1,4 @@
+'use client'
 import { useState } from 'react'
 
 import dynamic from 'next/dynamic'
